feat: auto approval permissions
