+++ e2e/configs/local/wdio.android.local.emu.conf.ts
@@ -16,6 +16,7 @@ config.capabilities = [
     'appium:app': resolve(__dirname, '../../apps', process.env.ANDROID_APP || 'BCWallet.apk'),
     'appium:noReset': process.env.NO_RESET === 'true',
     'appium:newCommandTimeout': 180,
+    'appium:autoGrantPermissions': true,
   },
 ]
