+++ e2e/configs/local/wdio.ios.local.device.conf.ts
@@ -22,6 +22,7 @@ config.capabilities = [
     'appium:showXcodeLog': process.env.SHOW_XCODE_LOG === 'true',
     'appium:noReset': process.env.NO_RESET === 'true',
     'appium:newCommandTimeout': 180,
+    'appium:autoAcceptAlerts': true,
   },
 ]
