+++ e2e/configs/local/wdio.ios.local.sim.conf.ts
@@ -16,6 +16,7 @@ config.capabilities = [
     'appium:app': resolve(__dirname, '../../apps', process.env.IOS_APP || 'BCWallet.app'),
     'appium:noReset': process.env.NO_RESET === 'true',
     'appium:newCommandTimeout': 180,
+    'appium:autoAcceptAlerts': true,
   },
 ]
