+++ e2e/configs/sauce/wdio.ios.sauce.rdc.conf.ts
@@ -14,6 +14,7 @@ config.capabilities = [
     'appium:noReset': false,
     'appium:fullReset': true,
     'appium:newCommandTimeout': 180,
+    'appium:autoAcceptAlerts': true,
     'sauce:options': {
       appiumVersion: 'latest',
       build: process.env.BUILD_NAME || `local-${Date.now()}`,
