+++ e2e/test/bcsc/onboarding/app-launch.spec.ts
@@ -1,5 +1,4 @@
 import { Timeouts } from '../../../src/constants.js'
-import { acceptLocalNetworkPermissionIfPresent } from '../../../src/helpers/iosPermissions.js'
 import { BaseScreen } from '../../../src/screens/BaseScreen.js'
 import { BCSC_TestIDs } from '../../../src/testIDs.js'
 
@@ -7,7 +6,6 @@ const AccountSetup = new BaseScreen(BCSC_TestIDs.AccountSetup)
 
 describe('App Launch', () => {
   it('should display the Account Setup screen', async () => {
-    await acceptLocalNetworkPermissionIfPresent()
     await AccountSetup.waitFor('AddAccount', Timeouts.appLaunch)
   })
 })
