+++ e2e/test/bcsc/onboarding/notifications.spec.ts
@@ -1,4 +1,3 @@
-import { acceptNotificationPermissionIfPresent } from '../../../src/helpers/notifications.js'
 import { BaseScreen } from '../../../src/screens/BaseScreen.js'
 import { BCSC_TestIDs } from '../../../src/testIDs.js'
 
@@ -8,6 +7,5 @@ describe('Notifications', () => {
   it('should navigate through the Notifications screen', async () => {
     await Notifications.waitFor('Continue')
     await Notifications.tap('Continue')
-    await acceptNotificationPermissionIfPresent()
   })
 })
